feat: Add comprehensive reproduction event and database schema checking utilities
- Implemented `check_reproduction_events()` to validate reproduction data across simulation databases
- Added robust schema checking with `check_db_schema()` function
- Enhanced logging and error handling for database and reproduction event analysis
- Included detailed checks for required columns and data integrity
- Improved debugging capabilities for database and reproduction event investigations
